add tests for header masking
