+++ Supabase/src/commonMain/kotlin/io/github/jan/supabase/exceptions/RestException.kt
@@ -1,11 +1,9 @@
 package io.github.jan.supabase.exceptions
 
-import io.github.jan.supabase.maskString
+import io.github.jan.supabase.maskHeaders
 import io.github.jan.supabase.maskUrl
 import io.ktor.client.statement.HttpResponse
 import io.ktor.client.statement.request
-import io.ktor.http.Headers
-import io.ktor.util.toMap
 
 /**
  * Base class for all response-related exceptions
@@ -32,16 +30,6 @@ open class RestException(val error: String, val description: String?, val respon
     val statusCode = response.status.value
 
 }
-
-private val SENSITIVE_HEADERS = listOf("apikey", "Authorization")
-
-private fun maskHeaders(headers: Headers): String = headers.toMap().mapValues { (key, value) ->
-    if(key in SENSITIVE_HEADERS) {
-        value.firstOrNull()?.let {
-            listOf(if(key == "Authorization") "Bearer ${maskString(it.drop(7), showLength = true)}" else maskString(it, showLength = true))
-        }
-    } else value
-}.toString()
 
 /**
  * Thrown when supabase-kt receives a response indicating an authentication error
